pyproject.toml
[tool.poetry]
name = "StockPulse"
version = "0.1.0"
description = ""
authors = ["Adriano R. de Sousa <adrianorodriguesdsousa@ugpi.edu.br>"]
readme = "README.md"

[tool.poetry.scripts]
server = "stockpulse.server:run"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pydantic = "1.9.0"
sqlmodel = "^0.0.12"
mysql-connector-python = "^8.2.0"
cryptography = "^41.0.5"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
